feat: restructure server & move packages, env config overrides, makefile, new tests
