+++ .env.example
@@ -1,0 +1,8 @@
+# Example environment configuration for dj-set-downloader
+
+# HTTP server
+SERVER_PORT=8000
+
+# Storage
+STORAGE_TYPE=local
+STORAGE_OUTPUT_DIR=output
