+++ internal/server/server.go
@@ -10,12 +10,14 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/jaki95/dj-set-downloader/config"
 	"github.com/jaki95/dj-set-downloader/internal/job"
+	"github.com/jaki95/dj-set-downloader/internal/service"
 )
 
 // Server handles HTTP requests for the DJ set processor
 type Server struct {
 	cfg        *config.Config
 	jobManager *job.Manager
+	processor  *service.Processor
 	router     *gin.Engine
 }
 
@@ -24,6 +26,7 @@ func New(cfg *config.Config) *Server {
 	return &Server{
 		cfg:        cfg,
 		jobManager: job.NewManager(),
+		processor:  service.NewProcessor(cfg),
 	}
 }
